refactor(send): split send amount auto-size input component
